pyproject.toml
[project]
name = "mcp-test"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.51.0",
    "beautifulsoup4>=4.13.4",
    "fastmcp>=2.3.4",
    "mcp[cli]>=1.8.1",
    "python-dotenv>=1.1.0",
]
